requirements.txt
Django==2.2.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
Pillow==6.1.0

